requirements.txt
kivy
python-dotenv
mysql-connector-python
